backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
google-auth
requests
pydantic
